requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
appdirs==1.4.3
argparse==1.4.0
astroid==1.5.2
atomicwrites==1.2.0
attrs==18.1.0
backports.functools-lru-cache==1.5
caniusepython3==7.0.0
certifi==2018.8.24
chardet==3.0.4
click-log==0.1.8
click==6.7
codecov==2.0.15
coverage==4.5.1
diff-cover==1.0.4
distlib==0.2.7
django==1.11.15
edx-i18n-tools==0.4.7
edx-lint==0.5.5
edx-opaque-keys==0.4.4
first==2.0.1
fs==2.0.18
future==0.16.0
idna==2.7
inflect==1.0.0            # via jinja2-pluralize
isort==4.3.4
jinja2-pluralize==0.3.0   # via diff-cover
jinja2==2.10              # via diff-cover, jinja2-pluralize
lazy-object-proxy==1.3.1
markupsafe==1.0           # via jinja2
mccabe==0.6.1
more-itertools==4.3.0
packaging==17.1
path.py==11.0.1           # via edx-i18n-tools
pbr==4.2.0
pip-tools==2.0.2
pluggy==0.7.1
polib==1.1.0              # via edx-i18n-tools
py==1.6.0
pycodestyle==2.4.0
pydocstyle==2.1.1
pygments==2.2.0           # via diff-cover
pylint-celery==0.3
pylint-django==0.7.2
pylint-plugin-utils==0.4
pylint==1.7.1
pymongo==3.7.1
pyparsing==2.2.0
pytest-cov==2.5.1
pytest-django==3.4.2
pytest==3.7.4
pytz==2018.5
pyyaml==3.13              # via edx-i18n-tools
requests==2.19.1
six==1.11.0
snowballstemmer==1.2.1
stevedore==1.29.0
tox-battery==0.5.1
tox==3.2.1
urllib3==1.23
virtualenv==16.0.0
wrapt==1.10.11
